Separate and improve test_PredictionsAndTanimotoScores
